chore: clear ruff 0.16 findings and pin ruff in CI (#207)
The lint job installs ruff unpinned (`pip install ruff`), so it tracks whatever
release is newest at run time. ruff 0.16.0 shipped on 2026-07-23 with an
expanded default rule set, and the existing tree tripped it 125 times — turning
the job red on `main` and on every open PR without a single commit having
touched the code. Verified before this change: ruff 0.16.0 reported exactly 125
errors on main and 125 on an unrelated translation branch.

Two halves, because either alone leaves the problem live:

Clears all 125 findings. The bulk is mechanical and was applied by ruff itself
(`--fix`, then `--unsafe-fixes` once each affected rule had been read):

  I001 (54) import blocks sorted; C408 (7) dict() calls to literals;
  RUF046 (3) int(round(x)) to round(x), which round() already returns;
  RUF059 (2) unused unpacked bindings underscored; UP031 (3) percent-format
  assertion messages to f-strings (still lazily evaluated by `assert`, so no
  behaviour change); SIM102 (4) nested guard ifs merged; SIM117 (30) nested
  `with` blocks combined.

Four needed judgement rather than a fix:

  - SIM222 flagged `assert "switch.close" in cover._pending_switch or True` —
    an assertion that cannot fail. Its own comment concedes the count may have
    been decremented by echo filtering, so the vacuous `or True` is dropped and
    the comment now states plainly that the test is a regression guard on the
    branch running, not on _pending_switch's contents.
  - S110/BLE001 on a try/except/pass in a test task-drainer become
    `contextlib.suppress(Exception)`, which is the same behaviour said outright.
  - PLW1510 (5) `subprocess.run` calls gain an explicit `check=False`. Every
    caller already asserts on `returncode`, so this documents what they do
    rather than changing it.

The 30 SIM117 sites are ones ruff declines to fix even unsafely (multi-line
context managers), so they were rewritten mechanically and checked two ways:
every `patch(...)` and `pytest.raises(...)` occurrence was counted before and
after per file and is unchanged, and the suite was re-run. An earlier attempt
at that rewrite silently dropped a `pytest.raises` from a triple-nested block;
tests/test_force_endpoint_redrive.py caught it, and the rewrite now only
collapses non-overlapping sites per parse.

Pins `ruff==0.16.0` in the workflow so a future release cannot do this again.
An unpinned linter makes CI non-reproducible: the same tree passes or fails
depending on the day. Bump it deliberately, in a PR that also clears whatever
the new version surfaces.

No production behaviour changes. 1387 Python tests and 387 frontend tests pass,
and ruff 0.16.0 reports a clean `check` and `format --check`.

+++ custom_components/cover_time_based/cover_wrapped.py
@@ -747,9 +747,9 @@ class WrappedCoverTimeBased(CoverTimeBased):
             if pos is not None:
                 self._log(
                     "_send_stop :: no native stop; freezing via set_cover_position(%d)",
-                    int(round(pos)),
+                    round(pos),
                 )
-                await self._call_set_cover_position(int(round(pos)))
+                await self._call_set_cover_position(round(pos))
                 return
         await self._call_cover_service("stop_cover")
